destory session before external login
combine destroy_existing_session and destroy_session

+++ lib/identity/helpers/auth.rb
@@ -264,13 +264,14 @@ module Identity::Helpers
     end
 
     def destroy_session
+      return if @cookie.session_id.nil?
       api = Identity::HerokuAPI.new(user: nil, pass: @cookie.access_token,
         ip: request.ip, request_ids: request_ids, version: 3)
       # tells API to destroy the session for Identity's current tokens, and
       # all the tokens that were provisioned through this session
       log :destroy_session, session_id: @cookie.session_id do
         api.delete(path: "/oauth/sessions/#{@cookie.session_id}",
-          expects: [200, 401])
+          expects: [200, 401, 404])
       end
     end
   end
